Add phonetic matching, ged_code_from_name Move multiple_replace, match_template_variables, earliest_index, get_matching_text_in_strs, get_matching_text, get_matching_then_nonmatching_text from hdx.utilities.text to hdx.utilities.matching Update pre-commit
