+++ requirements.txt
@@ -2,7 +2,7 @@
 #    uv pip compile pyproject.toml --resolver=backtracking --all-extras -o requirements.txt
 annotated-types==0.7.0
     # via pydantic
-attrs==24.2.0
+attrs==24.3.0
     # via
     #   frictionless
     #   jsonlines
@@ -10,25 +10,25 @@ attrs==24.2.0
     #   referencing
 beautifulsoup4==4.12.3
     # via hdx-python-utilities (pyproject.toml)
-certifi==2024.8.30
+certifi==2024.12.14
     # via requests
 cfgv==3.4.0
     # via pre-commit
 chardet==5.2.0
     # via frictionless
-charset-normalizer==3.3.2
+charset-normalizer==3.4.1
     # via requests
-click==8.1.7
+click==8.1.8
     # via typer
-coverage==7.6.1
+coverage==7.6.10
     # via pytest-cov
-distlib==0.3.8
+distlib==0.3.9
     # via virtualenv
-dnspython==2.6.1
+dnspython==2.7.0
     # via email-validator
 email-validator==2.2.0
     # via hdx-python-utilities (pyproject.toml)
-et-xmlfile==1.1.0
+et-xmlfile==2.0.0
     # via openpyxl
 filelock==3.16.1
     # via virtualenv
@@ -36,9 +36,9 @@ frictionless==5.18.0
     # via hdx-python-utilities (pyproject.toml)
 html5lib==1.1
     # via hdx-python-utilities (pyproject.toml)
-humanize==4.10.0
+humanize==4.11.0
     # via frictionless
-identify==2.6.1
+identify==2.6.5
     # via pre-commit
 idna==3.10
     # via
@@ -48,9 +48,9 @@ ijson==3.3.0
     # via hdx-python-utilities (pyproject.toml)
 iniconfig==2.0.0
     # via pytest
-isodate==0.6.1
+isodate==0.7.2
     # via frictionless
-jinja2==3.1.4
+jinja2==3.1.5
     # via frictionless
 jsonlines==4.0.0
     # via hdx-python-utilities (pyproject.toml)
@@ -58,9 +58,9 @@ jsonschema==4.23.0
     # via
     #   frictionless
     #   tableschema-to-template
-jsonschema-specifications==2023.12.1
+jsonschema-specifications==2024.10.1
     # via jsonschema
-loguru==0.7.2
+loguru==0.7.3
     # via
     #   hdx-python-utilities (pyproject.toml)
     #   pytest-loguru
@@ -68,7 +68,7 @@ markdown-it-py==3.0.0
     # via rich
 marko==2.1.2
     # via frictionless
-markupsafe==2.1.5
+markupsafe==3.0.2
     # via jinja2
 mdurl==0.1.2
     # via markdown-it-py
@@ -76,7 +76,7 @@ nodeenv==1.9.1
     # via pre-commit
 openpyxl==3.1.5
     # via hdx-python-utilities (pyproject.toml)
-packaging==24.1
+packaging==24.2
     # via pytest
 petl==1.7.15
     # via frictionless
@@ -84,19 +84,21 @@ platformdirs==4.3.6
     # via virtualenv
 pluggy==1.5.0
     # via pytest
-pre-commit==3.8.0
+pre-commit==4.0.1
     # via hdx-python-utilities (pyproject.toml)
-pydantic==2.9.2
+pydantic==2.10.4
     # via frictionless
-pydantic-core==2.23.4
+pydantic-core==2.27.2
     # via pydantic
-pygments==2.18.0
+pygments==2.19.1
     # via rich
-pytest==8.3.3
+pyphonetics==0.5.3
+    # via hdx-python-utilities (pyproject.toml)
+pytest==8.3.4
     # via
     #   hdx-python-utilities (pyproject.toml)
     #   pytest-cov
-pytest-cov==5.0.0
+pytest-cov==6.0.0
     # via hdx-python-utilities (pyproject.toml)
 pytest-loguru==0.4.0
     # via hdx-python-utilities (pyproject.toml)
@@ -125,24 +127,23 @@ requests-file==2.1.0
     # via hdx-python-utilities (pyproject.toml)
 rfc3986==2.0.0
     # via frictionless
-rich==13.8.1
+rich==13.9.4
     # via typer
-rpds-py==0.20.0
+rpds-py==0.22.3
     # via
     #   jsonschema
     #   referencing
-ruamel-yaml==0.18.6
+ruamel-yaml==0.18.10
     # via hdx-python-utilities (pyproject.toml)
-ruamel-yaml-clib==0.2.8
+ruamel-yaml-clib==0.2.12
     # via ruamel-yaml
 shellingham==1.5.4
     # via typer
-simpleeval==0.9.13
+simpleeval==1.0.3
     # via frictionless
-six==1.16.0
+six==1.17.0
     # via
     #   html5lib
-    #   isodate
     #   python-dateutil
 soupsieve==2.6
     # via beautifulsoup4
@@ -154,7 +155,7 @@ tabulate==0.9.0
     # via frictionless
 text-unidecode==1.3
     # via python-slugify
-typer==0.12.5
+typer==0.15.1
     # via frictionless
 typing-extensions==4.12.2
     # via
@@ -162,17 +163,19 @@ typing-extensions==4.12.2
     #   pydantic
     #   pydantic-core
     #   typer
-urllib3==2.2.3
+unidecode==1.3.8
+    # via pyphonetics
+urllib3==2.3.0
     # via requests
 validators==0.34.0
     # via frictionless
-virtualenv==20.26.6
+virtualenv==20.28.1
     # via pre-commit
 webencodings==0.5.1
     # via html5lib
 xlrd==2.0.1
     # via hdx-python-utilities (pyproject.toml)
-xlsx2csv==0.8.3
+xlsx2csv==0.8.4
     # via hdx-python-utilities (pyproject.toml)
 xlsxwriter==3.2.0
     # via tableschema-to-template
